+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - The `release_string_utf_chars` function has been marked as unsafe. (#334)
 - Breaking change: Mark `JNIEnv::new_direct_byte_buffer` as `unsafe` (#320)
+- The lifetime of `AutoArray` is no longer tied to the lifetime of a particular `JNIEnv` reference. (#302)
 
 ## [0.19.0] — 2021-01-24